old1/old/pyproject.toml
[project]
name = "task-4"
version = "0.1.0"
description = "A PDF Summarizer and Quiz Generator using OpenAI and Streamlit"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "openai>=1.3.3",
    "streamlit>=1.31.0",
    "python-dotenv>=0.21.0",
    "pypdf>=4.0.0"
]
